management: prepare dashboard data on the backend

+++ management-service/src/main/kotlin/cz/vutbr/fit/knot/enticing/management/managementservice/entity/ServerStatusEntity.kt
@@ -17,7 +17,7 @@ class ServerStatusEntity(
         @field:Id
         @field:JsonIgnore
         val id: Long,
-        val freePhysicalMemorySize: Long,
+        val freePhysicalMemorySize: Double,
         val processCpuLoad: Double,
         val systemCpuLoad: Double,
         val timestamp: LocalDateTime,
